marker cleanup
dangling markers on fm and am switching doesn't make much sense, fixed
to clear markers on switch. Considering the more complicated task of
clearing on reloads, or keeping marked ones on rezoom. right now it
seems sufficient if not optimal to require that the middle of the
contour stays in view when analyzing the contour